@@ -5,7 +5,7 @@
   "dependencies": {
     "@semantic-ui-react/css-patch": "^1.0.0",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^11.2.2",
+    "@testing-library/react": "^16.3.1",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.1",
     "axios": "^0.27.2",
@@ -13,13 +13,13 @@
     "history": "^5.3.0",
     "mobx": "^6.6.0",
     "mobx-react-lite": "^3.4.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.25.0",
     "react-scripts": "5.0.1",
     "react-toastify": "^9.1.3",
     "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^2.1.2",
+    "semantic-ui-react": "^2.1.5",
     "uuid": "^8.3.2",
     "web-vitals": "^2.1.4",
     "yup": "^1.7.1"
@@ -71,8 +71,8 @@
     "@types/prettier": "^3.0.0",
     "@types/q": "^1.5.8",
     "@types/range-parser": "^1.2.7",
-    "@types/react": "^17.0.1",
-    "@types/react-dom": "^17.0.1",
+    "@types/react": "^18.3.27",
+    "@types/react-dom": "^18.3.7",
     "@types/serve-index": "^1.9.4",
     "@types/trusted-types": "^2.0.7",
     "@types/uuid": "^8.3.4",
