@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import 'react-toastify/dist/ReactToastify.min.css';
 import './app/layout/styles.css';
 import App from './app/layout/App';
@@ -7,12 +7,14 @@ import reportWebVitals from './reportWebVitals';
 import { store, StoreContext } from './app/stores/store';
 import { ToastContainer } from 'react-toastify';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+
+root.render(
   <StoreContext.Provider value={store}>
     <ToastContainer position='bottom-right' hideProgressBar />
     <App />
-  </StoreContext.Provider>,
-  document.getElementById('root')
+  </StoreContext.Provider>
 );
 
 // If you want to start measuring performance in your app, pass a function
